service/requirements.txt
Django==2.0
django-filter==1.1.0
django-graphiql==0.4.4
djangorestframework==3.7.7
factory-boy==2.10.0
Faker==0.8.10
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
Markdown==2.6.11
promise==2.1
psycopg2==2.7.3.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
Rx==1.6.0
singledispatch==3.4.0.3
six==1.11.0
text-unidecode==1.1
typing==3.6.2
